refactor: typescript relicSetSolver
