+++ src/components/optimizerTab/OptimizerGrid.jsx
@@ -1,6 +1,6 @@
 import { AgGridReact } from 'ag-grid-react'
 import { defaultColDef, getBaseColumnDefs, getCombatColumnDefs, gridOptions } from 'components/optimizerTab/optimizerTabConstants.ts'
-import { OptimizerTabController } from 'lib/optimizerTabController.ts'
+import { OptimizerTabController } from 'lib/optimizerTabController'
 import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import { Flex, theme } from 'antd'
 import { arrowKeyGridNavigation } from 'lib/arrowKeyGridNavigation'
